IR: handle a different variant of run-as error
Currently, we decide that a device doesn't support run-as properly if its
output for a run-as command contains the string "run-as: Package '..' is
unknown". It turns out that in some cases the output could be:
  run-as: Could not set capabilities: Operation not permitted

This CL marks a device that reports that output to have a non-working
run-as as well.

Fixes https://code.google.com/p/android/issues/detail?id=219339

Change-Id: I13ec94788a41faaf97ea5ab02a05e802be7be7e8
(cherry picked from commit 1449dff25370db8b918e9ddbc114931c7739c02c)