Fix clear logcat button freezing logcat console output
This also should resolve the issue that some people reported very
slow clear logcat times. Especially this issue was reported only after
2.2, which corresponds to when we added the new AndroidLogcatService
feature.

The reason these issues were introduced was that AndroidLogcatView was
already maintaining an adb connection, and AndroidLogcatService started
managing it as well. When the user pressed clear, there was a chance
that AndroidLogcatService was reading from logcat at the same time
AndroidLogcatView cleared it. This triggers a platform bug present in
older versions (<= 22) of the platfrom that seems to hang adb logcat.

This is fixed by moving the clear logic into AndroidLogcatService,
which lets us explicitly restart the adb connection each time.

See https://android-review.googlesource.com/#/c/119673 for context.

Bug: http://b.android.com/211432
Bug: http://b.android.com/215392

Change-Id: I8437fd4e7363be3552af0efefb64abefc0f361f0