Address some BufferedImage leaks in Layout Editor
Most components are registered to be disposed with the project, meaning
they will never be eligible for GC (they will always be reachable from
the Disposer tree), and they will hold to expensive BufferedImages --
essentially the whole rendering history.

In this CL, we start to restrict the lifetimes of Nele components:
- NlEditorPanel instantiates a DesignSurface and NlModel, so these two
can be scoped by the NlEditor associated with this panel;
- NlPreviewForm creates a DesignSurface, and any number of NlModel
instances as requested; in this case we dispose of the NlModels
ourselves.

Bug: http://b.android.com/220120

Change-Id: I7555bc925ca5fd106e72ee9bf11b976365c57b39