Make the ModuleClassLoader prefer classes from the project
If a class is both present in the project and in the Android plugin, the
previous implementation would load the one in the Android plugin first.
This might cause some issues when they are our of sync.

Bug: http://b.android.com/220818

(cherry picked from commit bc4e78a92ecad58714bcea73b7f674bfb9432144)

and

Limit the mechanism to load classes from a project preferably to ones defined in Studio to only the solver classes.

Fixes: https://code.google.com/p/android/issues/detail?id=220818

Change-Id: I5f8d49ed3945e10c12b3f6aa5fb55ab99d36abad