Separated pre-sync cleanup tasks from validations.
Before this CL, PreSyncChecks was a monolithic class that over
time included a combination of pre-sync checks and project
settings cleanup. The code was not easy to follow and not very
testable.

With this CL, I splitted PreSyncChecks into smaller strategies
focused on doing one check only. This strategies are also
testable. I also moved the project settings cleanup tasks to
the class PreSyncProjectCleanUp.

This CL adds tests and to make code more testable, I converted
SdkSync and IdeSdks to full objects (instead of a collection of
static methods) and now they are IDE services that can be
easily mocked in tests.

Change-Id: I677f6f8ab0205f1ffdf1ab0991b7cb81834aa455