Remove 'gradle sync' methods from GradleProjectImporter. (CL 1 of 5.)
Before this CL, GradleProjectImporter had 2 functionalies:
1. Import a project into the IDE
2. Invoke Gradle sync for already imported projects
This class was already doing too much, and the code grow
to overcomplicated. For example, it had 3 overloaded
versions of 'importProject' and 3 overloaded versions
of 'requestSync', each of them with boolean parameters
that made client code hard to read.

This CL splits this functionality into 2 classes:
1. GradleProjectImporter: imports a project into the IDE
2. GradleSyncInvoker: triggers a Gradle sync for existing
   projects.
Each class has only 1 method, and the boolean parameters
have been consolidated in a 'request' class, with setters
for each boolean value, making the code self-documenting.

The code is simpler more testable, since collaborators can be
easily replaced my test doubles. It is also easier to
migrate away from IDEA's sync infrastrusture, since, with
this CL, we have 2 strategies for performing sync, the
old one using IDEA's code and the new one that we are
currently writing.

Change-Id: I92010308b3d635871871bcfed0d738e5809e4937