222354: Vector icons which use resources are not shown in the editor gutter
When vector images reference resources, they would not be properly rendered
in the editor (because the resource references were not resolved, so the
vector drawable rasterizer library would just throw exceptions, and we'd
fall back to the error icon).

This CL adds a pass where we resolve all the resources in the vector
(using the resource resolver) before attempting to rasterize it.

Some other improvements as well, and add unit test infrastructure for all
the resource annotations (as well as the new vector behavior).

Change-Id: I27f07d1a854c238cd5d6567cb4b2f23817a0422c