Wizard Config Project corrupted on a long company name
http://b.android.com/222373

The problem is with "Package name" text field. It uses a JBLabel,
and JBLabel's don't wrap text unless they contain html...
To fix the problem everywhere (including when creating a new Module)
LabelWithEditLink was re-written:
1 - Remove inner JPanel with CardLayout
2 - Remove the JBLabel
3 - Always show the text using JTextField
4 - Simplify toggle logic
5 - Simplify get/setText() and setFont()

Other nice side effect is that the height does not change (always displaying
the same JTextField), and it does not overflow to the link button.

Change-Id: I4e5e691b6fee6c7455ddda42b4dfa75e00ab525f