Add "offline" arg to gradlew in unit tests
AndroidGradleTestCase starts a gradlew process and allows it to
load test projects, but occassionally this has the side effect of
downloading files. We want to ensure that our unit tests do not
depend on having an internet connection to pass, so we disable that
ability, here, by executing gradle in offline mode.

Change-Id: I2014b0b3f95b6476894ca0fe2e611f74e849e11c