Fix gradle tests on Windows
This is a (sad) case of incorrect handling of "FILE:" URLs.

The "pathToUrl" method of the "com.intellij.openapi.vfs.VfsUtilCore"
class returns *invalid* "FILE:" URLs on Windows.

For example, in this case, we get this:

file://D:\src\studio-master-dev\tools\external\gradle\gradle-2.14.1-bin.zip

Note the "file://" prefix. The standard is either "file:///" or
"file:/".

Note that, on Linux, the method generates valid URLs because paths
always prefixed with "/". So, the URL for a similar file on Linux
would be

file:///src/studio-master-dev/tools/external/gradle/gradle-2.14.1-bin.zip

which is a valid URL.

So what happens with the invalid URL on Windows in this case
is that a "java.net.URISyntaxException" exception is thrown from the
"readDistroUrl" method of the "org\gradle\wrapper\WrapperExecutor"
class when Gradle tries to download the distribution file from the
local file system.

The fix is to use the build toURI().toURI() methods instead of
"pathToUrl".

Note that I started investigating fixing "pathToUrl" to generate
valid URLs in all case, but that function is not only used in many
places, there are also a lot of assumptions in the code base about
the particular format used by "pathToUrl", and finally these URLs
are sometimes saved into configuration files.

In essence, it would be a *much* bigger undertaking to fix how
IntelliJ deals with file URLs.

Change-Id: If144adbd1dfcb8e4d2ba58c8434f17d54ee21e52