Fixed bug where an exception thrown on project setup won't stop sync.
When an exception is thrown while setting up a project after sync,
the sync is marked as "failed", but the method
GradleSyncState#lastSyncFailed won't return 'true' because the
project will still have AndroidProjects. The result is that code
generation will still be executed, but it shouldn't.

This CL fixes this bug by setting 'setSyncErrorsFound' to 'true'
and adding a check to GradleSyncState#lastSyncFailed to verify
that there were sync errors. With this check, when an error
occurs while setting a project, sync will stop and won't
unnecessarily generate sources.

Change-Id: Ia49d95ff51933d20027d930af70978862ca8eb0d