Fixing incoherence of instrumentation runner class
This CL revert the revert of the CL http://ag/1541586 and fixes the remaing problems

The possibility of setting the instumentation runner class on the
running configuration of Android tests was available
(https://screenshot.googleplex.com/PeEODbWrD3W.png). Since it can be
defined in build.gradle file, there was a conflict between both
definitions, and then running the tests didn't work in these cases.

Since we want coherence between the tests when running from the command
line and from Android Studio, we don't want the user to be able to
define it in a place other than the gradle file.

This CL: removing possibility of setting it in the running configuration
when running a gradle project (https://screenshot.googleplex.com/i23Z9no6hy1.png)

Test
AndroidTestConfigurationProducerTest.testRunnerComponentsHiddenWhenGradleProject
is commented because of a memory leak in
AndroidRunConfigurationEditor<AndroidTestRunConfiguration>

Bug: http://b.android.com/224975
Change-Id: I4953b6d924756aef55b71f60a9f012bc5b43dce9