Disabled tests in DependencySetupTest that are failing in post-submit.
DependencySetupTest is *already* disabled in the Bazel PSQ, but the
problem is that the post-submit queue runs both Bazel and Ant builds.
DependencySetupTest is still running with Ant, but apparently the
prebuilt repo does not have needed dependencies yet, making the test
fail.

The quickest way to get the build green is to disable the test
methods. Then we can add the missing components to prebuilts and
enable the tests again.

Change-Id: I8457c7604ba7832aaa8441b519886c8c8f50201d