Expose settings to pick IR UI
This CL adds a combobox in the instant run settings to allow users to
pick amongst one of three UI options for instant run. The options are
creatively named "Option 1|2|3" in the UI because this is purely for a
UX test.

 - Option 1 is what is shipped right now (2.0 to 2.2)
 - Option 2 splits run into run and hotswap actions
 - Option 3 includes run (behaves similar to 1), and a new stop-and-run
   action which will do a full run (installs apk).

Options 2 and 3 are pretty similar - they are simply changing whether
the default action on run is to do an instant run or not.

Change-Id: I2667f7f9b25147820f9e7f10527cef1a35d437c3