Adding a public method to add a gradle update instruction.
If you create a Recipe as a node in another xml file, it will
be unmarshalled but will not contain the necessary instruction
that causes the gradle changes to be applied. Normally this is
added via the Recipe#parse call so this has been factored out
into a separate public method. Callers should reference this
in their afterUnmarshall method on the class containing the
Recipe instance.

Change-Id: I2e597b687b63b6d505e97ed54b22f29aa4a769da