Use GradleValues at some more places.
With this changes, all the return value (expect some in repositories
models) are Gradle Values.

Repositories models are returning default values when the repository
is not defined in the gradle file and that needs some special
handling to use the GradleValue there. I will do that in my next and
final change in this migration.

Change-Id: I5c86283f03dbb58bf313a7110d3b156a507150f0