Add IDE-side support for lint baseline reports
First, hook up baseline support to the Analysis action such that
when you run inspections in the IDE, issues are suppressed if
found in the baseline report.

Second, register the new baseline informational issue as an
IDE inspection such that when issues are suppressed, you get an
informational message in the inspection view mentioning that some
issues were hidden due to the baseline.

Also, gracefully handle new XML files (such as baseline reports)
outside the Android modules (for example at the root directory
level) without performing spell checking and bound-namespace checks
in them.

Change-Id: I5d1f7f11a208eef3eceb990ee4123a5191c0fbcc