Make it clear to the user that sync is completed even if populated from cache
The cache control path for gradle sync is followed often, including
Studio startup. Since in the beginning of sync a 'Gradle sync started'
event is issued anyway (we don't know yet whether it's going to be a
cache hit or not), we should make sure we post an event when sync is
logically complete, otherwise it is easy to think that sync is still
progressing even though it's not.

Change-Id: Ie25983d6ee336bbc61621aa464f52caab1ee629f