Fixed sync error handler to prevent duplicated exception messages.
This CL fixes a bug where the Gradle sync machinery, in the
case of a sync error, will create a ExternalSystemException passing
both the original message from the root cause, and the root cause.
This resulted in ExternalSystemException concatenating the
given message and the message from the cause. For example:

  "Failed to find Build Tools revision 24.0.3"

will end as

  "Failed to find Build Tools revision 24.0.3Failed to find Build Tools revision 24.0.3"

which is the cause for this bug.

Fixes:
https://code.google.com/p/android/issues/detail?id=226506

Test: ProjectImportErrorHandlerTest reproduces this bug and verifies that
the fix works.

Change-Id: I6193326de5a7d71695949143a1748c40b14e4183