Fix before launch tasks in JUnit run configuration
Currently, when a new before launch task is added to a JUnit run
configuration and after a Gradle Sync is run, this task is deleted and
the configuration returns to its default before launch tasks.

This CL fixes this problem by setting the before launch tasks in JUnit
run configurations correctly when doing a sync and adds a test for this
behaviour.

Test: PostSyncProjectSetupTest.testJUnitRunConfigurationSetup

Fixing bug: http://b.android.com/194704

Change-Id: Id4d853fdb7f0ced8bf085a9884355900a358ea75