Update Open File link to reference correct file
When a bad plugin dependency is found, search
the build files for the file that the plugin
dependency is specified in and show a link to
open that file.

This CL updates the AndroidPluginInfo class to
also return the file in which the plugin
dependency resides and allows it to run before
sync has been completed by no longer requiring
a valid Module to be found.

Test: OldAndroidPluginErrorHandlerTest now
checks the returned hyperlink references the
correct file and AndroidPluginInfoTest tests
that the correct plugin dependency location is
returned. the new PLUGIN_IN_APP project tests
the case where the plugin dependency is
specified directly in the app module's
gradle.build file instead of the project level
build file.

Fixes b.android.com/199419

Change-Id: I7ca7e997c2dafd7f88201bbfb2a9494b86d16ed7