Fix the bug that a process name remains [UNKNOWN].
StudioProfilerDeviceManager used to rely on deviceChanged() callback
to capture a newly deployed process. However, this callback is called
only when "Sent when a device data changed, or when clients are
started/terminated on the device."

Runtime is responsible to update a process's name but timing is the
key here. If the new app is deployed, deviceChanged() is called, and
then Runtime updates the process name, deviceChanged() won't be
called until another process is added or deleted which may not happen
very soon. Therefore, sometimes we see a process name remains [UNKNOWN].

The fix is clientChanged() callback.

I believe we may be able to refactor updateDevices() in the future
for performance (it may be called too often now).

Test: New feature; tests to follow.
Change-Id: I89156d58ce09b37362d8f17952e4c4f26a9fd2e2