First pass at documentation for the Assistant infrastructure
Quick first pass at documenting what the Assistant infra is good for
and how to use it.

FYI, the xml format is described rather than using a schema as it
is expected that parties will want to add their own custom dom.
We may want to revisit this in the future to optimize for returning
a config file rather than instance and help ensure it was properly
crafted.

Bug: 28022172

Test: NA, this is a documentation update.
Change-Id: Idf0d1ebc4aeae82c4502258216c53636f8a6aee0