"Declined upgrade" error message now shows the correct "recommended" plugin version.
This CL fixes the "recommended" plugin version in error
message displayed when the user declines to perform a
required plugin version upgrade.

The fix is quite simple, but testing was difficult since
the class PostSyncProjectSetup has too many responsibilities
and the code to test was inside private logic. This CL
refactors this logic into smaller classes that are testable.

Fixes:
https://code.google.com/p/android/issues/detail?id=227927

Test: ForcedPluginPreviewVersionUpgradeIdeaTest
Change-Id: I0b953e6eacf76432fb607a04b9dddff68a2c43eb