Fixes rendering of first and last samples in L1
Monitors requests the samples in the range (a,b), however
to render, for example, the memory level at time a, the samples
needed are the one before a and the one after a.

Given that this is an artifact of how the UI decides to
render, for example if we used stepped rendering, we wouldn't need an extra
sample at the end.

The change is localized to the UI. We could probably expand the API to include
how many samples are needed to the left and to the right of the request.
CPU could ask for an extra two samles to do the delta rendering too.

Test: Follow up
Change-Id: If480ee1d87d26827bdb52451d55cf17cb492b52b