Create Android JUnit configurations
In order to have more control over the the unit tests configurations,
some new RunConfiguration, RunConfigurationType and
RunConfigurationProducer are created based in JUnit plugin
implementations.

JUnit implementations are deactivated in AndroidStudio so we have only
one unit test configuration. When running in Idea, if it is an Android
Project, our RunConfigurations are preferred, and if not, the user can
choose which one he will use when running each test.

This CL allows future modifications in the code to be able to take into
account TestArtifactScopes when running Unit tests and also fixes some
bugs that otherwise would demand touching JUnit plugin code in a way
particular to Android's interests.

This is a first step in the direction of running tests through Gradle
Tooling API, since now we will have more control over the tests.

Test: tests were updated to take into account this new configurations
Change-Id: I64e825f2c30b564340fa41855fb1e77308ccffc1