Fixing scopes in JUnitConfiguration.Data
When an Android Instrumented Test class and a JUnit Test class have the
same package name and we try to run "all in package" unit tests, it
doesn't work. It happens because when trying to create the Unit Test
configuration it identifies all the test classes in the package and
tries to run the Android Test ones, failing.

This CL solves this problem by fixing the GlobalSearchScope used in
JUnitConfiguration.Data.

Test: JUnit test included

Fixes bug: http://b.android.com/217261

Change-Id: I593f79e5d5192fcd693bfb604a7e49ee4e3564a7