Revert "Revert Change Ied41ddee"
This reverts commit 71ee8c198a6be4e6e92ed88fc1d486cf7e2cafe8.

I think the cause of flakiness in this test is the method
AndroidGradleTestCase#refreshProjectRules, which is calling the refresh
method in LocalFileSystem non recursively, cleaning only the main
directory of the project but not sub directories. This CL attempts
to fix this problem.

Also, the dependency gson2.4 was not included in the dependencies of
the test data.This CL changes the dependency to gson2.2.4, that is
already included in the dependencies.

Test: same than before

Change-Id: Ie5ddf28ea07d19e92a5623e3699db0d3de09ade4