Fixes memory leak in AndroidRunConfigurationEditor
Memory leak problem is caused because all the JComponent objects aren't
Disposable but they have, for example, a strong reference to Project.
Also, AndroidRunConfigurationEditor is Disposable but never registered
in the Disposer.

There are two ways of fixing that: the first one is to register the
class in the Disposer, implement the method disposeEditor (override of
SettingsEditor's) by setting all the components as null; The second one
is by also registering the class in the Disposer but instead of
implementing the method, we avoid the strong reference in JBColor by
passing a lambda instead of an anonyme class.

This CL implements the second way.

Test: AndroidTestRunnerTest#testRunnerComponentsHiddenWhenGradleProject
reactivated.

Change-Id: Ifd8ba1e505b0d3fde01ca992475a9daa813a0055