ColumnTreeBuilder: Ensure tree always fills viewport
Before this change, if the tree contained in the ColumnTreeBuilder
component was smaller than the viewport (e.g. when the tree contains
very few elements), the tree height would be smaller than the viewport
height. Visually, this is not a problem, because they both share
the same background color. However, this is a problem wrt to the
"click-to-activate" behavior: If the mouse is clicked in the region
between the tree bottom and viewport bottom, the component is not
activate and does not receive the focus.

The change fixes that.

Test: Added a headless test that validates the new behavior.
Change-Id: I64fc4b78a2d93fca6fc55cbb5877fe68fd8ae577