Eliminate excessive error logging calls
IntelliJ logger implementation is such that each LOG.error(...) call
not only outputs the error information to the log file, but also
triggers a couple of UI events, which lead to two things:
1) ‘Internal IDE error’ balloon appearing
2) IDE event log in the UI gets populated with the error details

Except for that, errors impact crash/exception metrics which we
see on the analytics backend.

This all combined means that:
1) Multiple sequential calls to LOG.error() to indicate logically the
same error condition should never be used. Instead, the message details
should be accumulated in a StringBuffer and logged once.
2) It is worth considering LOG.error() replacement with LOG.warn()
in the cases when a dedicated error notification UI is implemented alongside
(e.g., Messages.showErrorDialog())
3) Although not noticed in the scope of this CL, usage of any logging
calls in a loop should be a subject to careful review

Test: This is a code health change which is
not testable per se. The sustainable way to keep this on the radar
is to document it in the team-wide style guide, so go/adtstyle
has been updated in that regard

Change-Id: I41b03adafb21562bae339ceb038a7a43ae6a6e9a