229643: Unable to add any constraints between UI controls
Fix class loading for AAR libraries. These don't ship with
R classes, but they're required at runtime. We compute these
on the fly, but the code which did this was relying on the
old exploded-aar directory layout to figure out the packages
to generate R classes for. This CL adjusts this scheme to
work with the build cache.

Test: This is covered by UI tests (once we update them to
use Gradle plugin 2.3 instead of 2.2.1.)

Change-Id: I578ec95ef212624a72d80a838318865872d8237d