Gracefully handle failures during split apk installs
If installing an APK fails, we capture the error and prompt the user if
they'd like to try again after uninstalling the app. This retry logic is
not provided for split apk installs.

This CL factors out the retry logic to a separate class
(RetyringInstaller), and reuses that from both the full and split apk
installers.

Fixes b.android.com/229769
Test: unit test added

Change-Id: I00a974f5d0072601d72697b593f5d1400ad7c992