Fix ExportSignedPackageTest and re-enable in Bazel
These tests had some problems as: it was referring direcly to testData
instead of creating the project in a tesmporary directory, versions of
gradle and gradle plugin were incorrect in the build files.

To fix these problems, now this class inherit from
AndroidGradleTestCase, which creates a temporary directory for the
project, updates all the versions in the build files and run the correct
version of Gradle.

The flakiness of this test should be fixed by now, since
AndroidGradleTestCase was fixed.

Test: fixing and re-enabling

Fixes bug: http://b.android.com/221311 and part of
http://b.android.com/226579

Change-Id: I8b7cbd0202afb5fcc0ba419898d8d2d1b577ee1a