Moves the concept of Time from the UI to the Model
This change changes the Choreographer to an Updater. This
new object lives in the model and now it's the model
itself that carries over time.

This changes the UI to be fully reactive and state-less,
having the following advantages:

* We play better with Swing, as our widgets can
  be used inside custom renderers.
* We don't render everything every frame anymore. Only
  components that receive the "update" have to be re-render.
  Using much less CPU if the UI is not scrolling.
* We can now unit-test in the model situations that
  require time passing.
* We can fix several glitches regarding time:
  . The scrollbar jumping around
  . The delay in animation when switching levels
  . The selection being not in sync in CPU
    profiling

Note this change is not complete yet, but current tests
still pass even though a few other visual glitches
are being introduced.

Change-Id: Ifafe1d2275c8784367abb33297fc2ef01286f9c8
Test: All existing tests pass