Update test configurations in AndroidCommonUtils
Since http://ag/1647523, the default JUnit configurations from junit
plugin have been replaced by new Android JUnit configurations so we
have more control over the code and we are able to modify scopes and
make Android Instrumented tests and JUnit tests work well together.

AndroidCommonUtils was not recognizing these new configurations and was
returning the wrong value in the method isTestConfiguration, which
caused an empty suite of tests error.

This CL fixes the problem by registering AndroidJUnitConfigurationType
in the list of known test types.

Test: AndroidJUnitConfigurationTypeTest guarantees that
isTestConfiguration works for this case.

Change-Id: Ia7a3eeb9b378b2db9731586c2e46c23c82300c78
Fixes: http://b.android.com/230419