Update render/inflate calls to be also async
To make the interface consistent, make the render and inflate calls also
async. Simply modified the callers to wait on the result right after
calling any of those methods so the behaviour should be roughtly the
same.
This won't have any immediate benefits but hopefully some of the callers
can be refactored to take advantage of the async behaviour.
Also updated the dispose call to wait for all the async tasks to finish
before proceeding to clean-up.

Test: Added new RenderTaskTest to exercise the render flow
Change-Id: I5084a034e12b27f7e1f9e420ad1572c22bc34b3a