Moves profiler android-plugin code to its own module
The new module profilers-android now contains all the code
that interacts with the profiler modules. This means that
the android module doesn't need to depend on the profiler ones.

Besides the obvious better abstraction, this means that
android_tests are no longer affected by changes
to profilers and profilers-ui. This should save 30 minutes
of PSQ time on these CLs.

Test: N/A
Bug: 33427831
Change-Id: I1b69d0d31b94d2863034f2410b9a25a7e8680576