Make http data readonly
We currently create a temporary file to store data
downloaded from a network request, but when the user
views it in its associated editor, it is sometimes
editable, which is weird.

Therefore, we mark the temp file as readonly, to ensure
that the view on the data is also read only.

Test: Tests updated to cover new code

Change-Id: Ic7cf70637181bc4e4e216795042f6e204ffed62c