Device Explorer Service interface and implementation
This is the first CL of a series of changes to implement the
Device Explorer window. This change contains only the lower level
layer, which is a thin abstraction over ADB file system services.

Upcoming changes will focus on the UI aspect of things.

Test: Simple start/restart of the ADB service. More coverage will be
possible with an upcoming change in ADB making it easier to write
tests against.

Change-Id: I93baa4757639851945e9d66d910159294dad17b3