Always check that gradle plugin is of a certain version for IR
At some point in 2.2, the gradle plugin provided an API to indicate whether it
supports instant run or not. Separately, the IR build-info.xml has a protocol
version, and if the IDE doesn't understand that protocol version, it'll stop
proceeding with the instant run build.

However, there is an issue with using 2.2 version of the gradle plugin (which
supports IR) with 2.3 of the IDE. In such a case, even though the plugin
supports IR, we don't want to enable IR.

So this CL adds in a version check where we make sure that gradle plugin is of
a certain version before querying for instant run status from it.

TODO: A subsequent CL will clean up the naming of the methods here..
Test: manual, unit tests coming in a subsequent CL
Fixes: b.android.com/231380

Change-Id: Ib0e73b103d871eb2017541f83d4530e02a58b133
(cherry picked from commit 41ea1035ae3098ebae8787418186fdabfb3cc624)