Fixed "forced" plugin upgrade when recommended version is a snapshot.
For example, if a project is using plugin 2.3.0-beta1 and the
recommended version is 2.3.0-dev, the IDE will no longer force
an upgrade.

As long as the recommended version is a snapshot (version ends
with "-dev") and the current is a preview of the same version
("-beta1"), the IDE will not prompt for a forced or recommended
plugin version update.

This fix is useful for testing with the dev version of the IDE.

Test: JUnit test.
Change-Id: I9b3bee4187dc88e1ecf5ebb344dfcc1b0dc72acb