Remove waitForBackgroundTasks after selecting device
The wait there is unnecessary, all the tests wait for some specific
condition anyway.  The timeout was occasionally hit when attaching the debugger.

Add log for whether stop button was actually clicked, which might be the
cause of some flaky failures and could have been made more frequent
with this change.

Test: N/A
Change-Id: Ia6efc7e607c9f8ef4e0e0a98ff027c1bf995b98d