Fix slider editing in the all properties table.
1)
When a user clicks on e.g. the elevation property a mousePressed
event is dispatched in the newly created editor. This works fine
for a normal text editor but when a slider is created the event
causes the slider value to change even through the user hasn't
seen the slider yet.
Fix this issue by ignoring mouse events in the slider until
a certain time elapse (0.3 secs) after the creation of the
control.

2)
After a slider value has changed, and the user is no longer
dragging the slider knob, we used to update the value in XML.
This caused the editor to disappear from the table.
Fix this by delaying the XML update to loss of focus in the
editor.

3)
Focus changes in the NlReferenceEditor was intercepted to
change the selection to highlight the field being edited.
Ignore focus changes if the focus is changed to/from the
slider.

4)
Handling of the escape key in the editor has been changed
to cause a cancelEditing table event instead of a
stopEditing event. (Found through testing).

Test: Unit tests have been added for NlReferenceEditor.
      A text fixture has been added.

Change-Id: I0c458a3f93e26884adc15948b5746a3fbf372a32