Automatic JUnit configurations migration
Since the new AndroidJUnit configurations were imlemented replacing and
improving the old JUnit configuraions, they were not migrated
automatically and then, if the user had a project with old JUnit
configurations, he would have to create manually the new ones.

This CL implements this migration automatically by implementing a
Converter and a ConverterProvider passing it. It reads the current
configurations, identifies which are JUnit and converts them.

Test: AndroidJUnitConfigurationConverter

Fixes bug http://b.android.com/230631

Change-Id: I32c8088061577871b1af6cb6adc7d16f67da4022