Wait for stop button to be showing before clicking
The stop button click did not wait for button.isShowing(), and when that
is not true the click action returns early.  Properly wait until the
button is showing before clicking.

Implemented as a separate method for now, I don't know the implications
of changing it for the other tests yet.  Eventually those other uses
should move to the new implementation, which also is has the more common
pattern of throwing if the button could not be clicked soon enough,
simplifying test code.

Test: N/A
Change-Id: Ic12f9a0e1ec34e49daa25cad4a20136ff78bb1ed