Faster render when zooming in Nele Editor
When zooming on the LayoutEditor, draw the Render result image in low quality and fast scaled.
The high quality image is drawn once the user stop zooming.

Also use the new draw method from the image pool to reduce the number
of Images created.

Test: Unit test using image comparaison to verify if the quality change
after changing the scale and repainting several times

Change-Id: I4363de7ae04e22ec7b171b454ebfbdb9ef07b678