ModelWizard handles exceptions on "Next" and "Finish" better now
Before, if the user hit "Next" and the next page threw an exception,
they would be left in a broken wizard. Now, the wizard keeps the user on
the same page and gives them a chance to recover.

Also, if the user hit "Finish" and one of the models throws an
exception, that would have been swallowed. Now, it is propogated
correctly to the user.

Test: ModelWizardTest tests added to cover new exception logic
Change-Id: If35776f71ac4d2be477dafffa7e6b08ba725b8e2