Change ModelWizard success boolean to enum, for clarity
Using "success" as a boolean for a wizard finishing was confusing, as one
of the wizard models may throw an exception during this step but the
wizard still has to close in that case. So, it doesn't really succeed,
but it's not cancelled, either, so we can't set "success" to false.

By making the state an explicit enum, the code is easier to read and
more honest about its intentions.

Test: Tests added to sanity check new enum type
Change-Id: Icdac059ad8f0ecf710701779c04e679eec693823