Don't say "No changes" for cold swap builds w/ no artifacts
The IDE notifies the user that there were no changes whenever gradle
does not deliver any artifacts to deploy. This is misleading in the
following scenario:
 - Perform a cold swap build.
 - Perform a number of hotswaps after that, with the last change undoing
   all the prior changes.
 - Now do a cold swap build.
At the last step, gradle compares the output from the prior cold swap
build and says that there is nothing to deploy, which is accurate, but
the user notification is wrong since there were changes since the
immediate prior hotswap.

This CL tweaks the notification to not mention "No changes" in such a
scenario (a cold swap + no artifacts).

Bug: http://b.android.com/232931
Test: included
Change-Id: Ib7f08a3fd2c6651032cd9c6fb97dcc24a33db25a
(cherry picked from commit fb73345ea7abbce58ba501279826edd99dfec560)