Profiler changed to poll processes/devices for connection.
Talking with Esteban we want to have the datastore query connections
for devices, and processes, this allows the datastore to know when
devices/processes are alive and enables us to keep this information
around for polling the devices / processes in the future.

This CL sets the framework for us supporting dead/past device queries,
the component missing is the start/end time being set for each device,
and process.

Test: ProfilerServiceTest
Change-Id: Ifb2e36178e468f076cc5f98bc13733bfc668f21e