Studio metrics: pass product version to UsageTracker
Right now we only track the version of Android Studio
doing the upload of the metrics. This can cause
really weird behavior if the user is running
two different versions of Android Studio Side-by-Side.
As the tracking and publishing are separated we can end up
 with:
-metrics reported for newer version
-metrics reported for older version

So instead we now track for each metric the version it was logged
as well as the version it was published with.

We can then process this on the server side.
Bug: http://b.android.com/232515
Test: Manually erified via "Android Studio Usage Tracker" internal UI

Change-Id: I4ae64c3c0291c841ebee6f6ec18e01f8bc458eee