Fix BottomUpNode calculation
Previously, the total time for a method was calculated
as the sum of all its durations, it works fine but not for a recursive
call.

For example:
main [0..40] -> B[1..40] -> B[2..40]

With previous calculation the total time for the method "B"
would be 39+38=77 and 77 is more than 40, which is wrong.

Current implementation calculates total time based on
union of ranges of the method.

Test: ag/1860981
Bug: b/35032367

Change-Id: I4262cc2761f3e9025bc42cd88d087eca3c7694a2