Fixed the way the IDE activates the "Project" view.
This CL fixes the way the IDE activates tool windows
by invoking the code the activation code inside
ToolWindowManager#invokeLater. This will ensure that
the tool window is ready by the time there is an
attempt to activate it.

The fix is only noticeable when activating the
"Project" view after importing an APK as a project.
Existing use cases were working fine.

Test: N/A
Change-Id: I7ca32c8439aca2bf1731a25c73afa6d22fa50769