Remove perModuleRepositories and mavenUrl from code
The per-module repository feature was not being used
and mavenUrl is no longer required for testing.

This CL removes code that is all on either branches
that were hard-coded to false, depended on external
values that are no longer set or are in classes
that are about to be removed.

Test: GradleImportTest and NewProjectTest (UI test)
Change-Id: I049dce59193465f6529b0d8c1f76ff4e11e9b458