Device Explorer: remove dependency on FileListingService
Still in preparation of supporting "run-as" and "su", the
AdbFileListing class now contains its own implementation of
executing "ls" commands on the target device to list files.

This removes the dependency on FileListingService from
AdbDeviceFileSystem. This does, however, preserves the same behavior
exposed to the API consumers, i.e. the Device Explorer UI.

This change also adds unit tests based on the output of a Nexus 7
device, ensuring the code that parses the output of "ls" commands is
covered.

Test: Existing unit tests + new ones.

Change-Id: Iebe589130cae3e80c5982cb1f3f693fe31885b7e