New Project Wizard: Missing attributes
Global attributes are missing when the project is created.
This attributes need to be available at the Project, Module and Render
levels.
Also removed a ugly cast (instanceof ModuleTemplateGalleryEntry) by
setting ATTR_INCLUDE_FORM_FACTOR on a different step

Test: Create a new Project with both Mobile and Wear and make sure
that the gradle file has a "wearApp project(...)", and that the Mobile
manifest has a WAKE_LOCK permission
Change-Id: I39b1d0681ed39f9b64c28e1ccb6479113e479384