Modules are now properly cleaned up if a Gradle models are not found.
The IDE now will clean up modules when Gradle models are not found
between sync operations. For example, if a module had a NDK facet,
but then the user removes support for NDK and syncs the project,
the module that once had NDK support should not have a NdkFacet after
sync.

For more details please read the bug report.

Fixes:
https://code.google.com/p/android/issues/detail?id=229806

Test: JUnit tests.
Change-Id: Ibe88382d37a87fdb8860a98eb21e33ca6eff00ee