Add a "slice" editor as an option for advanced property table.
Add a DOM style editor that only shows the attributes that have
a value assigned in the layout file. References to other xml files
are also shown with the values that are referenced.

This is currently hidden behind a flag such that only developer
builds will see the option in the UI.

Test: TBD.
Change-Id: I99cea6fa185ec29a9017d33b54b0e997e6e171de