Support Unix abstract sockets with JVMTI-enabled agents.
Unix abstract socket doesn't work between a shell user (perfd)
and an app user (agent). Therefore, perfd will first create a
client socket connecting to itself, then use execl() and run-as
to run a new process ('perfd --connect'). The new connector
process inherits the client socket's fd (thanks to excl()) and
is able to pass it to agent (thanks to run-as).

On Studio side, we need to call AttachAgent RPC to perfd.

Bug: 35672396

Test: manual
<enable.jvmti.profiling=true/false> X <O and older devices>

Change-Id: I0eb2313093b4bdf0b8e8c0eb4491cbfe9b752f68