Device Explorer: Detect shell commands capabilities
Older devices and API level don't support all shell commands
used by Device Explorer. This CL adds heuristics to detect
what commands are available (or not) and then use an alternative
command to achieve the same functionality. There are often
trade-offs when using alternative commands, so we still want to
use the "right" commands when they are available.

* Older devices don't support "rm -f", use plain "rm" instead

* Older devices don't have the "touch" command, use "echo -n >path"
  instead

* Older devices don't have the "test" command, use "ls -d -a" instead

Test: Updated existing tests, added tests for API 10 device

Change-Id: I9ea9ab1421468daf9977188f629bb4fab7a86e15