Fixed bug where "Gradle sync" action disappeared.
This happened right after a Gradle sync failure.
The source of the bug GradleProjectInfo#isBuildWithGradle.
After the latest merge with IDEA, this method cached
the returned value, and after a sync failure, the value
was never recomputed.

This CL fixes it by removing the cache. If the project
is a Gradle project, the method will return as soon as
it finds a module with the GradleModuleFacet, which is
not an expensive check.

Test: JUnit test.
Change-Id: I26c9817816144a655ba30d7c64170369874f4877