Track migration from ECJ to PSI
Lint now uses PSI directly instead of ECJ for Java parsing.

This moved some code from the android plugin up to lint
itself, so we can delete it here and reference the upstream
version instead.

Test: This code is covered by tests, and some of them are
updated in this CL to account for changes in how ECJ and
PSI treated identifier positions.

Change-Id: Icf46f9f5b1c070c6742dd7dbafae284ed8f32b30